+++ pytket/binders/passes.cpp
@@ -348,10 +348,15 @@ PYBIND11_MODULE(passes, m) {
       "When decomposing to CX, the substitution is only performed if it "
       "results in a reduction of the number of CX gates, or if at least "
       "one of the two-qubit gates is not a CX.\n\n"
+      "Using the `allow_swaps=True` (default) option, qubits will be "
+      "swapped when convenient to further reduce the two-qubit gate count "
+      "(only applicable when decomposing to CX gates).\n\n"
       ":param target_2qb_gate: OpType to decompose to. Either TK2 or CX.\n"
       ":param cx_fidelity: Estimated CX gate fidelity, used when "
-      "target_2qb_gate=CX.\n",
-      py::arg("target_2qb_gate") = OpType::CX, py::arg("cx_fidelity") = 1.);
+      "target_2qb_gate=CX.\n"
+      ":param allow_swaps: Whether to allow implicit wire swaps.",
+      py::arg("target_2qb_gate") = OpType::CX, py::arg("cx_fidelity") = 1.,
+      py::arg("allow_swaps") = true);
   m.def(
       "KAKDecomposition",
       [](double cx_fidelity) {
@@ -360,8 +365,8 @@ PYBIND11_MODULE(passes, m) {
       py::arg("cx_fidelity"));
   m.def(
       "DecomposeTK2",
-      [](const py::kwargs &kwargs) {
-        return DecomposeTK2(get_fidelities(kwargs));
+      [](bool allow_swaps, const py::kwargs &kwargs) {
+        return DecomposeTK2(get_fidelities(kwargs), allow_swaps);
       },
       "Decompose each TK2 gate into two-qubit gates."
       "\n\nGate fidelities can be passed as keyword arguments to perform "
@@ -377,11 +382,16 @@ PYBIND11_MODULE(passes, m) {
       "If no fidelities are provided, the TK2 gates will be decomposed "
       "exactly using CX gates.\n\n"
       "All TK2 gate parameters must be normalised, i.e. they must satisfy "
-      "`NormalisedTK2Predicate`."
-      "\n\nIf the TK2 angles are symbolic values, the decomposition will "
+      "`NormalisedTK2Predicate`.\n\n"
+      "Using the `allow_swaps=True` (default) option, qubits will be swapped "
+      "when convenient to reduce the two-qubit gate count of the decomposed "
+      "TK2.\n\n"
+      "If the TK2 angles are symbolic values, the decomposition will "
       "be exact (i.e. not noise-aware). It is not possible in general "
       "to obtain optimal decompositions for arbitrary symbolic parameters, "
-      "so consider substituting for concrete values if possible.");
+      "so consider substituting for concrete values if possible."
+      "\n\n:param allow_swaps: Whether to allow implicit wire swaps.",
+      py::arg("allow_swaps") = true);
   m.def(
       "NormaliseTK2", &NormaliseTK2,
       "Normalises all TK2 gates.\n\n"