+++ tket/src/Circuit/include/Circuit/CircPool.hpp
@@ -396,7 +396,8 @@ Circuit NPhasedX_using_PhasedX(
     unsigned int number_of_qubits, const Expr &alpha, const Expr &beta);
 
 /** TK2(a, b, c)-equivalent circuit, using normalised TK2 and single-qb gates */
-Circuit TK2_using_normalised_TK2(Expr ea, Expr eb, Expr ec);
+Circuit TK2_using_normalised_TK2(
+    const Expr &alpha, const Expr &beta, const Expr &gamma);
 
 // converts a TK1 gate to a PhasedXRz gate
 Circuit tk1_to_PhasedXRz(