+++ pytket/docs/changelog.rst
@@ -22,6 +22,8 @@ Minor new features:
   to accept classical controlled operations. ``ThreeQubitSquash`` and ``FullPeepholeOptimise``
   now accept classical operations.
 * Improve ``QControlBox`` decomposition.
+* New ``allow_swaps`` flag in ``KAKDecomposition`` and ``DecomposeTK2`` to
+  decompose two-qubit operations up to implicit wire swaps.
 
 Fixes: